refactor(eslint-plugin-lit-a11y): function->method
templateExpressionToHtml to TemplateAnalyzer
